requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
lxml==5.3.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypandoc==1.13
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validate-docbr==1.10.0
vine==5.1.0
wcwidth==0.2.13
